serpapi-google-mcp-server/pyproject.toml
[project]
name = "serpapi-google-mcp-server"
version = "0.1.0"
description = "A unified MCP server for SerpApi's Google data endpoints, providing tools for querying events, flights, jobs, hotels, maps, reviews, and news."
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["mcp>=1.6.0", "colorama==0.4.6", "httpx==0.28.1", "google-search-results==2.4.2", "python-dotenv==1.1.0"]

[[project.authors]]
name = "Rohit Ingole"
email = "rohit.vilas.ingole@gmail.com"

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project.scripts]
serpapi-google-mcp-server = "serpapi_google_mcp_server:main"
